feat: OPTIC-800: introducing stale feature flags (#5971)
### Describe the reason for change
We currently have many feature flags that OSS and on-prem customers can
toggle. This is acceptable in a limited number of scenarios if any, but
over time, these code branches become obsolete. We are investing ongoing
effort, that will continue after this PR, to minimize these dead code
branches.


To accelerate this process, we introduce the concept of stale feature
flags. We will gradually populate this list and clean the code for these
feature branches of code accordingly.

The list of FF to stale was categorized considering:
-  Default values in SaaS
-  Default values in on-prem
-  Default values in Open Source

It involved a manual check followed by a second pass comparing it with
the current feature_flag.json file.

We will keep adding new FF to this stale list and remove the associated
code afterward. This also provides time to identify any mistakes.

Notice this cover access to FF on JS and Python.


#### Change has impacts in these area(s)
Frontend Feature Flags consumption
Backend Feature Flags consumption

#### What does this fix?

This will help maintain a healthier codebase over time.


#### What is the new behavior?

For this initial set of FF, we were very restrictive, aiming for minimal
impact.



#### What alternative approaches were there?

We could use an array that we don't allow to modify, but we aim to make
this change more self-documented. This way, the fixed values are more
easily readable in the code.

---------

Co-authored-by: Raul Martin <raul@humansignal.com>
Former-commit-id: 2185f721a6e3d907feb491d6a3a8f993c295e947