Experiments to add region indices to audio regions
Audio has a very different approach to normal react app,
incapsulating on changes inside in internal regions,
not reacting on outer store changes and not providing tools to sync.
Some approaches were made in this commit to react on different changes
in the store and change internal accordingly.
Failed.

Also the text cut-off is fixed in this commit.


Former-commit-id: e0faa51ad80909695c767fa3d1a589873c2c7acf