feat: LEAP-1441: Allow to connect comments to classifications (#6486)
Modify comments to allow to attach results. This includes global classifications like Choices and TextArea + per-region results connected to a region.

### Code
Technically this adds `controlName` to `Anchor` to distinguish regions from results. Bboxes are created over the tag defined by this `controlName`. Comment can be attached to existing result only, so if tag has no value added to it, it can't be attached.

### UI
In linking mode there is a highlight around a tag holding a result, so user can select it to attach the result. Any change to result is reflected in a comment. If that's the per-region result it will be styled as a region and will display region index. Comment icons are added directly to a result.


Former-commit-id: 8484a8489a41ca698fb02777904b84e655f7a79f