fix: LEAP-1581: Improve styles in new comments (#6537)
There is a handy `-webkit-line-clamp` rule! All items are inlined to be just a text, but truncated after 3 lines, so should be good for a lot of different content.
Also scales down the comment icon to always fit it inside usual editor screen with paddings. Fixes LEAP-1578
Style fixes for comment's classification after LEAP-1476

Former-commit-id: 02eb60246d28113a45a6a975cf7ac84901e99ea8