fix: LEAP-1335: Visually disable MIG navigation in View All (#6702)
Currently View All mode doesn't support any kind of interactions due to the way it's rendered. So pagination for MIG images won't work. We need to make it more clear for users, visually disable buttons and add a tooltip to indicate this.

Our tooltips are also a kind of interaction, so we use simple html `title`.

Co-authored-by: MihajloHoma <MihajloHoma@users.noreply.github.com>
Former-commit-id: d0084d26138857e27e4545e42d7c0e5c1f4707e8